Better leverage existing invalid generators
Generating random strings only to check if they conform to a condition
is inefficient. Besides, writing such conditions seemingly leads to
duplicate work (when already writing a parser and valid generators).
Ultimately, the conditions here would have turned into "parsers"
anyway.

In a case as with expressions (e.g., where an invalid expression is
**neither** a valid literal nor a valid name), it is better to begin
with some invalid generator, and check that the generated string is
**not otherwise valid** (e.g., an invalid name that is not otherwise a
valid literal).